Move tips code to TipsView
