+++ app/templates/navbars/navbar.html.j2
@@ -60,7 +60,7 @@
           </a>
 
           <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-            <a class="dropdown-item" href="{{ url_for('SupportView:tips') }}"> Tipy a triky </a>
+            <a class="dropdown-item" href="{{ url_for('TipsView:index') }}"> Tipy a triky </a>
           </div>
         </li>
